backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
pypdf
sentence-transformers
pinecone-client
google-generativeai
